backend/api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "FastAPI microservice"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "strawberry-graphql>=0.288.2",
]
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
